Add solution and fix bug in the terminating condition in while loop
When using two pointers, we should always check if fast.next is not
null to avoid the edge case when we only have two nodes in the list.
If we only check `while(fast)`, we will get an error "cannot read
property next null" telling us that fast is null.